fix: added ruff to uv packages and solved linting issues

+++ services/orchestrator/src/orchestrator/utils/compute_percentile.py
@@ -7,5 +7,5 @@ def percentile(values: list[int], p: float) -> float | None:
     if len(sorted_values) == 1:
         return float(sorted_values[0])
 
-    index = int(round((len(sorted_values) - 1) * p))
+    index = round((len(sorted_values) - 1) * p)
     return float(sorted_values[index])
